\name{eucl_dist}
\alias{eucl_dist}

\title{
		Computes the euclidean distance between the rows of two different matrices
}

\description{
	This functions is used internally by essR to compute the euclidean distance between the rows of two different matrices. 
	The matrices must have the same number of columns.
}


\note{
	For internal use of MEIGOR.
}