\name{MEIGOR-package}
\alias{MEIGOR-package}
\docType{package}

\title{
	Global optimization Toolbox
}

\description{
	A global optimization package containing several algorithms such as a scatter search implementation and
	variable neighborhood search (plus cooperative multicore/multimachine implentations of these) 
	and dynamic hill climbing.
}

\details{
	\tabular{ll}{
		Package: \tab MEIGOR\cr
		Type: \tab Package\cr
		Version: \tab 0.99.0\cr
		Date: \tab 2012-09-10\cr
		License: \tab GPLv3\cr
		LazyLoad: \tab yes\cr
	}
}

\author{

	Jose Egea
	David Henriques
	Thomas Cokelaer
	Alejandro F. Villaverde
	Julio R. Banga
	Julio Saez-Rodriguez

	Maintainer:Jose Egea <josea.egea@upct.es>
	
}
\references{
 Egea, J.A., Maria, R., Banga, J.R. (2010) An evolutionary method for
 complex-process optimization. Computers & Operations Research 37(2):
 315-324.
 
 Egea, J.A., Balsa-Canto, E., Garcia, M.S.G., Banga, J.R. (2009) Dynamic
 optimization of nonlinear processes with an enhanced scatter search
 method. Industrial & Engineering Chemistry Research 49(9): 4388-4401. 
}
\keyword{ package }
\keyword{scatter}
\keyword{search}
\keyword{optimization}

\seealso{
	\code{\link{essR}}
	\code{\link{essR}}
}
