\name{optim_fobj}
\alias{optim_fobj}

\title{
	Gateway function to evaluate the objective function when the local solvers are invoked.
}

\description{
	This function is used internally by essR to evaluate the objective function when the local solvers are invoked.
}

\note{
	For internal use of MEIGOR.
}