\name{solnp_eq}

\alias{solnp_eq}

\title{
	Gateway function to evaluate the equality constraints when solnp is invoked as local solver
}
\description{
	This function is used by essR to evaluate the equality constraints when solnp is invoked as local solver
}
\note{
	For internal use of MEIGOR.
}

