\name{solnp_fobj}

\alias{solnp_fobj}


\title{
	Gateway function to evaluate the objective function when solnp is invoked as local solver
}

\description{
	This functions is used internally by essR to evaluate the objective function when solnp is invoked as local solver
}

\note{
	For internal use of MEIGOR.
}