\name{solnp_ineq}

\alias{solnp_ineq}

\title{
	Gateway function to evaluate the inequality constraints when solnp is invoked as local solver
}

\description{
	This function is used internally by essR to evaluate the inequality constraints when solnp is invoked as local solver
}

\note{
	For internal use of MEIGOR.
}
