\name{ssm_beyond}

\alias{ssm_beyond}

\title{
	Function that expands the search direction when a good offspring solution has been found
}

\description{
	This function is used internally by essR to expand the search direction when a good offspring solution has been found
}

\note{
	For internal use of MEIGOR.
}