\name{ssm_penalty_function}

\alias{ssm_penalty_function}

\title{
	Calculates the penalized objective function in constrained problems
}

\description{
	This function is used internally by essR to calculate the penalized objective function in constrained problems
}

\note{
	For internal use of MEIGOR.
}